backend/pyproject.toml
[tool.poetry]
name = "openfyxer-backend"
version = "1.0.0"
description = "OpenFyxer - Self-hosted AI Executive Assistant Backend"
authors = ["OpenFyxer Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"
websockets = "^12.0"

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
aiosqlite = "^0.19.0"

# Neo4j
neo4j = "^5.16.0"

# Redis
redis = "^5.0.1"
celery = {extras = ["redis"], version = "^5.3.6"}

# Authentication & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyotp = "^2.9.0"
qrcode = {extras = ["pil"], version = "^7.4.2"}
cryptography = "^42.0.0"

# Validation
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"

# Email Integration
google-api-python-client = "^2.114.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
msal = "^1.26.0"
imapclient = "^2.3.1"
aiosmtplib = "^3.0.1"

# Calendar Integration
icalendar = "^5.0.11"

# LLM Integration
openai = "^1.10.0"
anthropic = "^0.14.0"
google-generativeai = "^0.3.2"
cohere = "^4.44"
llama-cpp-python = "^0.2.32"
httpx = "^0.26.0"

# RAG & Embeddings
llama-index = "^0.9.40"
llama-index-llms-ollama = "^0.1.2"
llama-index-graph-stores-neo4j = "^0.1.3"
llama-index-embeddings-huggingface = "^0.1.4"
sentence-transformers = "^2.3.1"

# Document Processing
pypdf = "^3.17.4"
python-docx = "^1.1.0"
beautifulsoup4 = "^4.12.3"
html2text = "^2020.1.16"

# Speech-to-Text
faster-whisper = "^0.10.0"

# Utilities
python-dateutil = "^2.8.2"
pytz = "^2024.1"
langdetect = "^1.0.9"
tenacity = "^8.2.3"
aiofiles = "^23.2.1"

# HTTP Client
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
